requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pdfplumber>=0.9.0
pillow>=10.0.0
pandas>=2.0.0
python-docx>=1.0.0
plotly>=5.0.0
langchain
langchain-community
langchain-core
langchain-text-splitters
chromadb
